+++ runtime/binding-asyncapi/src/main/java/io/aklivity/zilla/runtime/binding/asyncapi/internal/config/AsyncapiNamespaceGenerator.java
@@ -14,17 +14,18 @@
  */
 package io.aklivity.zilla.runtime.binding.asyncapi.internal.config;
 
-import static com.fasterxml.jackson.dataformat.yaml.YAMLGenerator.Feature.MINIMIZE_QUOTES;
-import static com.fasterxml.jackson.dataformat.yaml.YAMLGenerator.Feature.WRITE_DOC_START_MARKER;
 import static io.aklivity.zilla.runtime.common.feature.FeatureFilter.featureEnabled;
 import static java.util.stream.Collectors.toList;
 import static org.agrona.LangUtil.rethrowUnchecked;
 
 import java.io.StringReader;
 import java.io.StringWriter;
+import java.time.Duration;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
+import java.util.Optional;
 import java.util.function.ToLongFunction;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
@@ -33,20 +34,17 @@ import java.util.stream.Collectors;
 import jakarta.json.Json;
 import jakarta.json.JsonObject;
 import jakarta.json.JsonReader;
+import jakarta.json.JsonValue;
 import jakarta.json.JsonWriter;
 import jakarta.json.bind.Jsonb;
 import jakarta.json.bind.JsonbBuilder;
 
 import org.agrona.collections.MutableInteger;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.dataformat.yaml.YAMLMapper;
-
 import io.aklivity.zilla.runtime.binding.asyncapi.config.AsyncapiOptionsConfig;
 import io.aklivity.zilla.runtime.binding.asyncapi.config.AsyncapiServerConfig;
 import io.aklivity.zilla.runtime.binding.asyncapi.internal.model.Asyncapi;
+import io.aklivity.zilla.runtime.binding.asyncapi.internal.model.AsyncapiKafkaServerBindings;
 import io.aklivity.zilla.runtime.binding.asyncapi.internal.model.AsyncapiSchemaItem;
 import io.aklivity.zilla.runtime.binding.asyncapi.internal.model.AsyncapiServer;
 import io.aklivity.zilla.runtime.binding.asyncapi.internal.model.AsyncapiTrait;
@@ -54,6 +52,7 @@ import io.aklivity.zilla.runtime.binding.asyncapi.internal.view.AsyncapiSchemaVi
 import io.aklivity.zilla.runtime.binding.asyncapi.internal.view.AsyncapiServerView;
 import io.aklivity.zilla.runtime.catalog.inline.config.InlineOptionsConfig;
 import io.aklivity.zilla.runtime.catalog.inline.config.InlineSchemaConfigBuilder;
+import io.aklivity.zilla.runtime.catalog.karapace.config.KarapaceOptionsConfig;
 import io.aklivity.zilla.runtime.engine.config.BindingConfig;
 import io.aklivity.zilla.runtime.engine.config.BindingConfigBuilder;
 import io.aklivity.zilla.runtime.engine.config.MetricRefConfig;
@@ -63,8 +62,9 @@ import io.aklivity.zilla.runtime.engine.config.TelemetryRefConfigBuilder;
 
 public abstract class AsyncapiNamespaceGenerator
 {
-    protected static final String INLINE_CATALOG_NAME = "catalog0";
+    protected static final String CATALOG_NAME = "catalog0";
     protected static final String INLINE_CATALOG_TYPE = "inline";
+    protected static final String KARAPACE_CATALOG_TYPE = "karapace";
     protected static final String VERSION_LATEST = "latest";
     protected static final AsyncapiOptionsConfig EMPTY_OPTION =
         new AsyncapiOptionsConfig(null, null, null, null, null, null, null);
@@ -218,13 +218,33 @@ public abstract class AsyncapiNamespaceGenerator
         NamespaceConfigBuilder<C> namespace,
         List<Asyncapi> asyncapis)
     {
-        final boolean injectCatalog = asyncapis.stream()
-            .anyMatch(a -> a.components != null && a.components.schemas != null && !a.components.schemas.isEmpty());
-        if (injectCatalog)
+        Optional<AsyncapiServer> server = asyncapis.stream()
+            .filter(a -> a.servers.entrySet().stream().anyMatch(s ->  s.getValue().bindings != null &&
+                s.getValue().bindings.kafka != null))
+            .map(s -> s.servers.values())
+            .flatMap(Collection::stream)
+            .filter(s -> s.bindings.kafka != null)
+            .findFirst();
+        final boolean injectCatalog = asyncapis.stream().anyMatch(AsyncapiNamespaceGenerator::hasSchemas);
+        if (server.isPresent())
+        {
+            AsyncapiKafkaServerBindings kafka = server.get().bindings.kafka;
+            namespace
+                .catalog()
+                    .name(CATALOG_NAME)
+                    .type(KARAPACE_CATALOG_TYPE)
+                    .options(KarapaceOptionsConfig::builder)
+                        .url(kafka.schemaRegistryUrl)
+                        .context("default")
+                        .maxAge(Duration.ofHours(1))
+                        .build()
+                    .build();
+        }
+        else if (injectCatalog)
         {
             namespace
                 .catalog()
-                    .name(INLINE_CATALOG_NAME)
+                    .name(CATALOG_NAME)
                     .type(INLINE_CATALOG_TYPE)
                     .options(InlineOptionsConfig::builder)
                         .subjects()
@@ -242,14 +262,10 @@ public abstract class AsyncapiNamespaceGenerator
     {
         for (Asyncapi asyncapi : asyncapis)
         {
-            if (asyncapi.components != null && asyncapi.components.schemas != null && !asyncapi.components.schemas.isEmpty())
+            if (hasSchemas(asyncapi))
             {
                 try (Jsonb jsonb = JsonbBuilder.create())
                 {
-                    YAMLMapper yaml = YAMLMapper.builder()
-                        .disable(WRITE_DOC_START_MARKER)
-                        .enable(MINIMIZE_QUOTES)
-                        .build();
                     for (Map.Entry<String, AsyncapiSchemaItem> entry : asyncapi.components.schemas.entrySet())
                     {
                         AsyncapiSchemaView schema = AsyncapiSchemaView.of(asyncapi.components.schemas, entry.getValue());
@@ -257,7 +273,7 @@ public abstract class AsyncapiNamespaceGenerator
                         subjects
                             .subject(entry.getKey())
                             .version(VERSION_LATEST)
-                            .schema(writeSchemaYaml(jsonb, yaml, schema))
+                            .schema(writeSchemaJson(jsonb, schema))
                             .build();
                     }
                     if (asyncapi.components.messageTraits != null)
@@ -268,7 +284,7 @@ public abstract class AsyncapiNamespaceGenerator
                                 subjects
                                     .subject(k)
                                     .version(VERSION_LATEST)
-                                    .schema(writeSchemaYaml(jsonb, yaml, v))
+                                    .schema(writeSchemaJson(jsonb, v))
                                     .build());
                         }
                     }
@@ -282,39 +298,40 @@ public abstract class AsyncapiNamespaceGenerator
         return subjects;
     }
 
-    protected static String writeSchemaYaml(
+    private static boolean hasSchemas(
+        Asyncapi asyncapi)
+    {
+        return asyncapi.components != null &&
+            asyncapi.components.schemas != null &&
+            !asyncapi.components.schemas.isEmpty();
+    }
+
+    protected static String writeSchemaJson(
         Jsonb jsonb,
-        YAMLMapper yaml,
         Object schema)
     {
-        String result = null;
-        try
+        String schemaJson = jsonb.toJson(schema);
+
+        JsonReader reader = Json.createReader(new StringReader(schemaJson));
+        JsonValue jsonValue = reader.readValue();
+
+        if (jsonValue instanceof JsonObject)
         {
-            String schemaJson = jsonb.toJson(schema);
+            JsonObject jsonObject = (JsonObject) jsonValue;
 
-            JsonReader reader = Json.createReader(new StringReader(schemaJson));
-            JsonObject jsonObject = reader.readObject();
-
-            JsonObject modifiedJsonObject = jsonObject.getJsonObject("schema");
-
-            if (modifiedJsonObject != null)
+            if (jsonObject.containsKey("schema"))
             {
+                JsonValue modifiedJsonValue = jsonObject.get("schema");
                 StringWriter stringWriter = new StringWriter();
                 JsonWriter jsonWriter = Json.createWriter(stringWriter);
-                jsonWriter.writeObject(modifiedJsonObject);
+                jsonWriter.write(modifiedJsonValue);
                 jsonWriter.close();
 
                 schemaJson = stringWriter.toString();
             }
+        }
 
-            JsonNode json = new ObjectMapper().readTree(schemaJson);
-            result = yaml.writeValueAsString(json);
-        }
-        catch (JsonProcessingException ex)
-        {
-            rethrowUnchecked(ex);
-        }
-        return result;
+        return schemaJson;
     }
 
     protected  <C> BindingConfigBuilder<C> injectMetrics(